vmc/vio_module/requirements.txt
cython
loguru==0.5.3
paho-mqtt==1.5.1
transforms3d
numpy
colored
pyrealsense